Moved test in Pattern only class
